pyproject.toml
[tool.poetry]
name = "indors-test-task"
version = "0.1.0"
description = "Test task on Indors vacansy"
authors = ["Aleksandr Rol <rolaleksandr75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
webcolors = "^24.8.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.8.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
